feat: move json_response and concat_into_url methods to http_tools
