+++ src/tractusx_sdk/dataspace/adapters/connector/base_dma_adapter.py
@@ -21,6 +21,7 @@
 #################################################################################
 
 from ..adapter import Adapter
+from ...tools import HttpTools
 
 
 class BaseDmaAdapter(Adapter):
@@ -29,7 +30,7 @@ class BaseDmaAdapter(Adapter):
     def __init__(self, base_url: str, dma_path: str, headers: dict = None):
         self.dma_path = dma_path
 
-        dma_url = self.concat_into_url(base_url, dma_path)
+        dma_url = HttpTools.concat_into_url(base_url, dma_path)
         super().__init__(dma_url, headers)
 
     class _Builder(Adapter._Builder):
